fix: move files to src
